# openapi.json
{"openapi":"3.0.0","paths":{"/kpi":{"post":{"operationId":"KpiController_createKPI","summary":"Register a KPI","description":"Create a KPI for display","parameters":[{"name":"x-api-key","in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateKpi"}}}},"responses":{"201":{"description":"Successful request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KPI"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unauthorized"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Conflict"}}}}}},"patch":{"operationId":"KpiController_updateKPI","summary":"Update a KPI","description":"Update a KPI for display","parameters":[{"name":"x-api-key","in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateKpi"}}}},"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KPI"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unauthorized"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Conflict"}}}}}},"delete":{"operationId":"KpiController_deleteKPI","summary":"Delete a KPI","description":"Delete a KPI","parameters":[{"name":"id","required":true,"in":"query","example":2,"description":"Id of the KPI to delete","schema":{"type":"number"}},{"name":"x-api-key","in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"type":"boolean"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unauthorized"}}}}}},"get":{"operationId":"KpiController_getKPI","summary":"Display a KPI","description":"Display a KPI","parameters":[{"name":"id","required":true,"in":"query","example":2,"description":"Id of the KPI to display","schema":{"type":"number"}},{"name":"x-api-key","in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KPI"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unauthorized"}}}}}}},"/kpi/list":{"get":{"operationId":"KpiController_getKPIList","summary":"Display a KPI List","description":"Display the KPI List for the given dates and type","parameters":[{"name":"from","required":true,"in":"query","description":"The start date in ISO8601 format","example":"2020-02-23T09:28:47.081Z","schema":{"format":"date-time","type":"string"}},{"name":"to","required":true,"in":"query","description":"The end date in ISO8601 format","example":"2023-02-23T09:28:47.081Z","schema":{"format":"date-time","type":"string"}},{"name":"type","required":true,"in":"query","description":"Type of the KPI to display","schema":{"enum":["success_rate__transfer_orange_ci","success_rate__transfer_mtn_ci","success_rate__transfer_moov_ci","success_rate__transfer_wave_ci","success_rate__payment_orange_ci","success_rate__payment_mtn_ci","success_rate__payment_moov_ci","success_rate__payment_wave_ci","success_rate__transfer_all","success_rate__payment_all","volume_xaf","volume_xof","transaction_last_one"],"type":"string"}},{"name":"x-api-key","in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/KPI"}}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unauthorized"}}}}}}},"/kpi/latest":{"get":{"operationId":"KpiController_getLatestKPI","summary":"Display mean values for all KPI","description":"Display mean values for all registered KPI since 1970.","parameters":[{"name":"Accept-Language","in":"header","description":"Language used to format response data","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KpiDataDto"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}}}}},"/kpi/private":{"get":{"operationId":"KpiController_getPrivateKPI","summary":"Display private KPIs","description":"Display private KPIs.","parameters":[{"name":"Accept-Language","in":"header","description":"Language used to format response data","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KpiDataDto"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}}}}},"/kpi/raw":{"get":{"operationId":"KpiController_getKPIRaw","summary":"Display mean values for all KPI","description":"Display mean values for all registered KPI since 1970. If a KPI is NotFoundException\n    displayed, it means it has not been registered yet","parameters":[],"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KpiDataDto"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}}}}},"/kpi-info":{"post":{"operationId":"KpiInfoController_createKpiInfo","summary":"Register KPI infos","description":"Register title and description for a KPI ","parameters":[{"name":"x-api-key","in":"header","schema":{"type":"string"}},{"name":"Accept-Language","in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KpiInfoDto"}}}},"responses":{"201":{"description":"Successful request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KpiInfoDto"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unauthorized"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Conflict"}}}}}},"get":{"operationId":"KpiInfoController_getKPIInfo","summary":"Display a KPI info","description":"Display a KPI info","parameters":[{"name":"type","required":true,"in":"query","example":"transaction_last_one","description":"Type of the KPI info to display","schema":{"type":"string"}},{"name":"Accept-Language","in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KpiInfo"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}}}}},"/kpi-info/{id}":{"patch":{"operationId":"KpiInfoController_updateKPIInfo","summary":"Update a KPI infos","description":"Update a KPI info","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"x-api-key","in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateKpiInfoDto"}}}},"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KpiInfo"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unauthorized"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Conflict"}}}}}},"delete":{"operationId":"KpiInfoController_deleteKpiInfo","summary":"Delete a KPI info","description":"Delete a KPI info","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"x-api-key","in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"Successful request","content":{"application/json":{"schema":{"type":"boolean"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BadRequest"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Unauthorized"}}}}}}}},"info":{"title":"Test API","description":"Description","version":"1.0","contact":{}},"tags":[{"name":"Your Tag","description":""}],"servers":[],"components":{"schemas":{"CreateKpi":{"type":"object","properties":{"date":{"type":"string","description":"Date of the KPI","example":"2023-02-23T09:28:47.081Z"},"type":{"type":"string","description":"Type of the KPI","example":"volume_xof","enum":["success_rate__transfer_orange_ci","success_rate__transfer_mtn_ci","success_rate__transfer_moov_ci","success_rate__transfer_wave_ci","success_rate__payment_orange_ci","success_rate__payment_mtn_ci","success_rate__payment_moov_ci","success_rate__payment_wave_ci","success_rate__transfer_all","success_rate__payment_all","volume_xaf","volume_xof","transaction_last_one"]},"value":{"type":"object","description":"Value of the KPI","example":{"value":2}}},"required":["date","type","value"]},"KPI":{"type":"object","properties":{"id":{"type":"number","description":"Automatic generated ID","example":1},"date":{"type":"string","description":"Date of the KPI","example":"2023-02-23T09:28:47.081Z"},"type":{"type":"string","description":"Type of the KPI","enum":["success_rate__transfer_orange_ci","success_rate__transfer_mtn_ci","success_rate__transfer_moov_ci","success_rate__transfer_wave_ci","success_rate__payment_orange_ci","success_rate__payment_mtn_ci","success_rate__payment_moov_ci","success_rate__payment_wave_ci","success_rate__transfer_all","success_rate__payment_all","volume_xaf","volume_xof","transaction_last_one"],"example":"volume_xof"},"value":{"type":"object","description":"Data for that KPI","example":{"value":2}}},"required":["id","date","type","value"]},"BadRequest":{"type":"object","properties":{"error":{"type":"string","description":"Error name","example":"Bad Request"},"message":{"type":"string","description":"Error message","example":"date must be a valid ISO 8601 date string"},"statusCode":{"type":"number","description":"Error code","example":400}},"required":["error","message","statusCode"]},"Conflict":{"type":"object","properties":{"error":{"type":"string","description":"Error name","example":"Conflict exception"},"statusCode":{"type":"number","description":"Error code","example":409}},"required":["error","statusCode"]},"Unauthorized":{"type":"object","properties":{"error":{"type":"string","description":"Error name","example":"Unauthorized"},"statusCode":{"type":"number","description":"Error code","example":401}},"required":["error","statusCode"]},"UpdateKpi":{"type":"object","properties":{"id":{"type":"number","description":"Id of the KPI to update","example":2},"date":{"type":"string","description":"Date of the KPI","example":"2023-02-23T09:28:47.081Z"},"type":{"type":"string","description":"Type of the KPI","example":"volume_xof","enum":["success_rate__transfer_orange_ci","success_rate__transfer_mtn_ci","success_rate__transfer_moov_ci","success_rate__transfer_wave_ci","success_rate__payment_orange_ci","success_rate__payment_mtn_ci","success_rate__payment_moov_ci","success_rate__payment_wave_ci","success_rate__transfer_all","success_rate__payment_all","volume_xaf","volume_xof","transaction_last_one"]},"value":{"type":"object","description":"Value of the KPI","example":{"value":2}}},"required":["id","date","type","value"]},"KpiDataDto":{"type":"object","properties":{"startDate":{"type":"string","description":"Start date for which KPIS are computed","example":"2018-02-23T09:28:47.081Z"},"endDate":{"type":"string","description":"End date for which KPIS are computed","example":"2023-02-23T09:28:47.081Z"},"kpi":{"type":"object","description":"Mean values of the different KPIS for the given dates","example":{"success_rate__transfer_wave_ci":91.25,"success_rate__transfer_orange_ci":95}}},"required":["startDate","endDate","kpi"]},"KpiInfoDto":{"type":"object","properties":{"type":{"type":"string","description":"Type of the KPI","enum":["success_rate__transfer_orange_ci","success_rate__transfer_mtn_ci","success_rate__transfer_moov_ci","success_rate__transfer_wave_ci","success_rate__payment_orange_ci","success_rate__payment_mtn_ci","success_rate__payment_moov_ci","success_rate__payment_wave_ci","success_rate__transfer_all","success_rate__payment_all","volume_xaf","volume_xof","transaction_last_one"],"example":"volume_xof"},"title":{"type":"string","description":"Title","example":"Success Rate - PAY-OUT Orange CI"},"description":{"type":"string","description":"Description","example":"This is the success rate for PAY-OUT using the Orange CI provider"}},"required":["type","title","description"]},"UpdateKpiInfoDto":{"type":"object","properties":{"type":{"type":"string","description":"Type of the KPI","enum":["success_rate__transfer_orange_ci","success_rate__transfer_mtn_ci","success_rate__transfer_moov_ci","success_rate__transfer_wave_ci","success_rate__payment_orange_ci","success_rate__payment_mtn_ci","success_rate__payment_moov_ci","success_rate__payment_wave_ci","success_rate__transfer_all","success_rate__payment_all","volume_xaf","volume_xof","transaction_last_one"],"example":"success_rate__transfer_orange_ci"},"title":{"type":"string","description":"Title","example":"Success Rate - PAY-OUT Orange CI"},"description":{"type":"string","description":"Description","example":"This is the success rate for PAY-OUT using the Orange CI provider"}},"required":["type","title","description"]},"KpiInfo":{"type":"object","properties":{"id":{"type":"number","description":"Automatic generated ID","example":1},"type":{"type":"string","description":"Type of the KPI","enum":["success_rate__transfer_orange_ci","success_rate__transfer_mtn_ci","success_rate__transfer_moov_ci","success_rate__transfer_wave_ci","success_rate__payment_orange_ci","success_rate__payment_mtn_ci","success_rate__payment_moov_ci","success_rate__payment_wave_ci","success_rate__transfer_all","success_rate__payment_all","volume_xaf","volume_xof","transaction_last_one"],"example":"volume_xof"},"title":{"type":"string","description":"Title","example":"Success Rate - PAY-OUT Orange CI"},"description":{"type":"string","description":"Description","example":"This is the success rate for PAY-OUT using the Orange CI provider"},"language":{"type":"string","description":"Language","example":"en"}},"required":["id","type","title","description","language"]}}}}